LayoutTests/fast/forms/search/search-recent-searches-expected.txt

PASS recentSearches.length is 1
PASS recentSearches[0] is "searchValue"
PASS successfullyParsed is true

TEST COMPLETE
